+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/editor/holders/text/Description.kt
@@ -5,6 +5,7 @@ import android.widget.TextView
 import com.anytypeio.anytype.core_ui.features.editor.BlockViewHolder
 import com.anytypeio.anytype.presentation.editor.editor.model.BlockView
 import kotlinx.android.synthetic.main.item_block_description.view.*
+import timber.log.Timber
 
 class Description(view: View) : BlockViewHolder(view) {
 
@@ -12,8 +13,12 @@ class Description(view: View) : BlockViewHolder(view) {
 
     fun bind(view: BlockView.Description) {
         if (view.mode == BlockView.Mode.READ) enableReadMode() else enableEditMode()
-        content.pauseTextWatchers {
-            content.setText(view.description, TextView.BufferType.EDITABLE)
+        if (!content.hasFocus()) {
+            content.pauseTextWatchers {
+                content.setText(view.description, TextView.BufferType.EDITABLE)
+            }
+        } else {
+            Timber.d("Skipping binding for block in focus")
         }
     }
     fun enableReadMode() {